requirements.txt
google-genai>=1.0.0
anthropic>=0.40.0
manim>=0.18.0
python-dotenv>=1.0.0
rich>=13.0.0
gradio>=4.0.0
mcp>=1.0.0
elevenlabs>=1.0.0
moviepy>=1.0.3
pydub>=0.25.1
arxiv>=2.0.0
supabase>=2.0.0

